Bring back upstream support for no_std in Android
As the upstream crate was designed to be no_std (therefore uses core)
and as the Soong targets need to build a dylib (which requires std), the
crate source was being patched to remove the no_std crate attribute.

In turn, as the crate doesn't define a Rust edition, Cargo (and
therefore cargo_embargo) defaults to 2015, which requires explicitly
importing the core crate in no-no_std environments so the code was
further patched to use std modules in place of the corresponding core
modules.

Instead, patch the Cargo.toml to use Rust 2018 (which implicitly imports
core, even in no-no_std) and wrap the no_std in a AOSP-only cfg that we
enable when building Soong targets that include dylib variants.

Enable that cfg from the Android.bp & cargo_embargo.json.

Bug: 324074084
Test: external_updater/updater.sh update --force rust/crates/memoffset
Test: cargo_embargo generate cargo_embargo.json
Test: m libmemoffset libintrusive_collections
Change-Id: Id76df918c0ec89651c4b84402967570e52ae1058